# Making the AVX-512 UTF-8 to UTF-16 transcoder more branchy