In this PR (to be applied to https://github.com/simdutf/simdutf/pull/97), we add a few more branches to help the performance of some Asian inputs. The net result is a slight loss of performance all around, with a nice gain in the few cases where our haswell kernel was slower. This might prove to be an interesting compromise since with this new code, the icelake kernel is always faster than haswell, except maybe on emoji inputs. 

The gist of the idea is to check when our expanded 512-bit registers contain too few valid characters and to merge two successive expanded 512-bit registers when we detect such a scenario. When this branch is not taken, we pay for some arithmetic and a branch not taken. However, when the branch is fruitful, the gains are interesting:  from   3.3 GB/s to 3.9 GB/s over some Japanese file (a 15% boost).

```
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 69840, iterations: 10000, dataset: unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   3.851 GB/s (1.3 %)    1.294 Gc/s     2.98 byte/char
convert_utf8_to_utf16+icelake, input size: 69840, iterations: 10000, dataset: unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   3.331 GB/s (1.4 %)    1.119 Gc/s     2.98 byte/char
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 67808, iterations: 10000, dataset: unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   3.578 GB/s (1.3 %)    1.233 Gc/s     2.90 byte/char
convert_utf8_to_utf16+icelake, input size: 67808, iterations: 10000, dataset: unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   3.338 GB/s (0.9 %)    1.151 Gc/s     2.90 byte/char
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 66600, iterations: 10000, dataset: unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   2.206 GB/s (1.1 %)    0.899 Gc/s     2.45 byte/char
convert_utf8_to_utf16+icelake, input size: 66600, iterations: 10000, dataset: unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   3.340 GB/s (0.9 %)    1.361 Gc/s     2.45 byte/char
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
```



After:


```
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 69840, iterations: 10000, dataset: unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   3.838 GB/s (1.6 %)    1.289 Gc/s     2.98 byte/char
convert_utf8_to_utf16+icelake, input size: 69840, iterations: 2000, dataset: unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   3.955 GB/s (0.8 %)    1.329 Gc/s     2.98 byte/char 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 67808, iterations: 10000, dataset: unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   3.587 GB/s (1.5 %)    1.236 Gc/s     2.90 byte/char
convert_utf8_to_utf16+icelake, input size: 67808, iterations: 2000, dataset: unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   3.932 GB/s (0.9 %)    1.355 Gc/s     2.90 byte/char 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
Using ICU version 70.1
testcases: 1
input detected as UTF8
current system detected as icelake
===========================
convert_utf8_to_utf16+haswell, input size: 66600, iterations: 10000, dataset: unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   2.204 GB/s (3.4 %)    0.898 Gc/s     2.45 byte/char
convert_utf8_to_utf16+icelake, input size: 66600, iterations: 2000, dataset: unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   3.946 GB/s (0.9 %)    1.608 Gc/s     2.45 byte/char 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).

```