@lemire 2022-07-15
With this PR, the icelake kernel beats the haswell kernel effectively everytimes for "convert_utf8_to_utf16".

```
$ for i in unicode_lipsum/lipsum/*.utf8.txt; do echo $i; (./buildnewnew/benchmarks/benchmark -F $i -P convert_utf8_to_utf16+icelake | grep GB ) && (./buildnewnew/benchmarks/benchmark -F $i -P convert_utf8_to_utf16+haswell | grep GB ); done
unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
   3.358 GB/s (0.7 %)    1.881 Gc/s     1.78 byte/char 
   2.388 GB/s (1.1 %)    1.338 Gc/s     1.78 byte/char 
unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   3.961 GB/s (1.1 %)    1.331 Gc/s     2.98 byte/char 
   3.832 GB/s (1.2 %)    1.287 Gc/s     2.98 byte/char 
unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
   3.013 GB/s (0.7 %)    0.753 Gc/s     4.00 byte/char 
   1.580 GB/s (1.2 %)    0.395 Gc/s     4.00 byte/char 
unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
   3.347 GB/s (1.0 %)    1.878 Gc/s     1.78 byte/char 
   2.387 GB/s (0.8 %)    1.339 Gc/s     1.78 byte/char 
unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
   3.980 GB/s (13.6 %)    1.482 Gc/s     2.69 byte/char 
   2.571 GB/s (1.3 %)    0.957 Gc/s     2.69 byte/char 
unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   3.964 GB/s (0.8 %)    1.366 Gc/s     2.90 byte/char 
   3.549 GB/s (1.3 %)    1.223 Gc/s     2.90 byte/char 
unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   3.965 GB/s (0.8 %)    1.616 Gc/s     2.45 byte/char 
   2.196 GB/s (1.2 %)    0.895 Gc/s     2.45 byte/char 
unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  20.985 GB/s (5.1 %)   20.985 Gc/s     1.00 byte/char 
  21.899 GB/s (3.0 %)   21.899 Gc/s     1.00 byte/char 
unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
   3.329 GB/s (0.8 %)    1.842 Gc/s     1.81 byte/char 
   2.476 GB/s (1.0 %)    1.370 Gc/s     1.81 byte/char 
```

```
$ for i in unicode_lipsum/wikipedia_mars/*.utf8.txt; do echo $i; (./buildnewnew/benchmarks/benchmark -F $i -P convert_utf8_to_utf16+icelake | grep GB ) && (./buildnewnew/benchmarks/benchmark -F $i -P convert_utf8_to_utf16+haswell | grep GB ); done
unicode_lipsum/wikipedia_mars/arabic.utf8.txt
   4.164 GB/s (0.9 %)    3.314 Gc/s     1.26 byte/char 
   3.169 GB/s (4.1 %)    2.522 Gc/s     1.26 byte/char 
unicode_lipsum/wikipedia_mars/chinese.utf8.txt
   4.037 GB/s (0.8 %)    3.055 Gc/s     1.32 byte/char 
   3.017 GB/s (3.3 %)    2.283 Gc/s     1.32 byte/char 
unicode_lipsum/wikipedia_mars/czech.utf8.txt
   3.938 GB/s (0.7 %)    3.709 Gc/s     1.06 byte/char 
   3.022 GB/s (2.8 %)    2.846 Gc/s     1.06 byte/char 
unicode_lipsum/wikipedia_mars/english.utf8.txt
  13.367 GB/s (2.3 %)   13.269 Gc/s     1.01 byte/char 
  13.002 GB/s (2.3 %)   12.907 Gc/s     1.01 byte/char 
unicode_lipsum/wikipedia_mars/esperanto.utf8.txt
   5.892 GB/s (1.2 %)    5.700 Gc/s     1.03 byte/char 
   5.318 GB/s (1.3 %)    5.145 Gc/s     1.03 byte/char 
unicode_lipsum/wikipedia_mars/french.utf8.txt
   4.439 GB/s (1.3 %)    4.320 Gc/s     1.03 byte/char 
   2.914 GB/s (4.4 %)    2.835 Gc/s     1.03 byte/char 
unicode_lipsum/wikipedia_mars/german.utf8.txt
   6.029 GB/s (1.1 %)    5.896 Gc/s     1.02 byte/char 
   5.478 GB/s (1.1 %)    5.357 Gc/s     1.02 byte/char 
unicode_lipsum/wikipedia_mars/greek.utf8.txt
   4.106 GB/s (0.8 %)    3.238 Gc/s     1.27 byte/char 
   3.399 GB/s (0.9 %)    2.681 Gc/s     1.27 byte/char 
unicode_lipsum/wikipedia_mars/hebrew.utf8.txt
   3.677 GB/s (0.7 %)    2.831 Gc/s     1.30 byte/char 
   2.892 GB/s (0.9 %)    2.226 Gc/s     1.30 byte/char 
unicode_lipsum/wikipedia_mars/hindi.utf8.txt
   4.665 GB/s (1.0 %)    3.222 Gc/s     1.45 byte/char 
   2.527 GB/s (4.3 %)    1.745 Gc/s     1.45 byte/char 
unicode_lipsum/wikipedia_mars/japanese.utf8.txt
   4.138 GB/s (2.6 %)    2.993 Gc/s     1.38 byte/char 
   3.308 GB/s (1.2 %)    2.393 Gc/s     1.38 byte/char 
unicode_lipsum/wikipedia_mars/korean.utf8.txt
   3.973 GB/s (0.9 %)    2.960 Gc/s     1.34 byte/char 
   2.744 GB/s (1.0 %)    2.045 Gc/s     1.34 byte/char 
unicode_lipsum/wikipedia_mars/persan.utf8.txt
   3.969 GB/s (0.7 %)    3.168 Gc/s     1.25 byte/char 
   3.130 GB/s (0.9 %)    2.498 Gc/s     1.25 byte/char 
unicode_lipsum/wikipedia_mars/portuguese.utf8.txt
   4.745 GB/s (0.8 %)    4.626 Gc/s     1.03 byte/char 
   3.798 GB/s (1.9 %)    3.703 Gc/s     1.03 byte/char 
unicode_lipsum/wikipedia_mars/russian.utf8.txt
   3.811 GB/s (0.8 %)    2.921 Gc/s     1.30 byte/char 
   2.784 GB/s (2.2 %)    2.134 Gc/s     1.30 byte/char 
unicode_lipsum/wikipedia_mars/thai.utf8.txt
   4.783 GB/s (1.1 %)    3.262 Gc/s     1.47 byte/char 
   2.882 GB/s (2.8 %)    1.965 Gc/s     1.47 byte/char 
unicode_lipsum/wikipedia_mars/turkish.utf8.txt
   4.682 GB/s (0.8 %)    4.451 Gc/s     1.05 byte/char 
   3.555 GB/s (1.3 %)    3.380 Gc/s     1.05 byte/char 
unicode_lipsum/wikipedia_mars/vietnamese.utf8.txt
   3.884 GB/s (0.9 %)    3.438 Gc/s     1.13 byte/char 
   1.676 GB/s (2.5 %)    1.483 Gc/s     1.13 byte/char 
```