@clausecker 2022-07-15
Looking good!  Still not quite 2x Haswell, but it's a lot better than before.  Meanwhile, I've also made good inroads into the problem. Will let you know once it works.  I have added some spiffy new ideas to make it go faster.