@lemire 2022-07-15
@clausecker I am sure that we can do much better but I wanted to make sure that I caught the low-hanging fruits.