@lemire 2022-07-17
This has been merged into avx512feature.