@clausecker 2022-07-27
Are these performance numbers with validation?  I've started to benchmark my own code and found that it is unfortunately not faster than this one (yet).  Will have to fiddle with the code and see if I can find improvements (very likely):

```
8to16LE/avx512/../unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt       3.58GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt      4.60GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt        4.34GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt       3.54GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt        4.02GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt     4.34GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt       3.99GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt        19.6GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt      3.60GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/arabic.utf8.txt      3.88GB/s ± 1%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/chinese.utf8.txt     3.36GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/czech.utf8.txt       2.94GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/english.utf8.txt     12.6GB/s ± 4%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/esperanto.utf8.txt   4.69GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/french.utf8.txt      3.52GB/s ± 1%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/german.utf8.txt      4.83GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/greek.utf8.txt       3.80GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/hebrew.utf8.txt      3.18GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/hindi.utf8.txt       4.33GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/japanese.utf8.txt    3.51GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/korean.utf8.txt      3.34GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/persan.utf8.txt      3.55GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/portuguese.utf8.txt  3.81GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/russian.utf8.txt     3.60GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/thai.utf8.txt        4.76GB/s ± 1%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/turkish.utf8.txt     3.52GB/s ± 0%
8to16LE/avx512/../unicode_lipsum/wikipedia_mars/vietnamese.utf8.txt  3.04GB/s ± 0%
```